+++ crates/weaver_semconv/src/v2/attribute_group.rs
@@ -89,6 +89,7 @@ impl AttributeGroup {
                     entity_associations: vec![],
                     visibility: Some(AttributeGroupVisibilitySpec::Internal),
                     is_v2: true,
+                    span_name_note: None,
                 }
             }
             AttributeGroup::Public(public) => {
@@ -122,6 +123,7 @@ impl AttributeGroup {
                     entity_associations: vec![],
                     visibility: Some(AttributeGroupVisibilitySpec::Public),
                     is_v2: true,
+                    span_name_note: None,
                 }
             }
         }